Correct issues with the unix version of libiked. Fix bugs that were causing the bidirectional message functions from working correctly. Also rename open and close to attatch and detach so we avoid namespace conflicts with the socket library functions.
Correct issues with the unix ikec program. QT requires that events are used so that gui elements are only updated by the 
main application thread. Use events instead of manipulating the elements directly in the ikec connect thread. Also 
ignore sigpipe errors so the application doesn't bomb when iked disconnects from the client.


git-svn-id: svn://svn.shrew.net/ike/head@72 1bb156eb-19af-db11-b919-0014c23ffdf2