Correct a typo in ikec that was preventing DNS settings from being forwarded to iked. Add basic support to iked for writing settings to the resolv.conf file. An attempt will be made to move an existing resolv.conf file to resolv.iked when a tunnel is connected and then restore the file after the tunnel is closed.
Modify iked to run in a daemon mode by default and accept the -F parameter to run in a foregroung mode. A few other command line options still need to be supported.

Modify iked to attempt to load the if_tap kernel driver when initialized. This is helpful for situations where the kernel does not have support compiled in.


git-svn-id: svn://svn.shrew.net/ike/head@78 1bb156eb-19af-db11-b919-0014c23ffdf2