Modify the way we use divert rules on win32. We now dynamically insert and remove these rules as required. This allows the Shrew Soft client to peacefully co-exist with other software that uses intermediate drivers such as the Cisco VPN Client.
Cleanup the packet dump functions and variables to be more self descriptive. Win32 registry values have also been modified to match.

Change the packet filter matching option to not be special cased for NAT-T NON-ESP packets. Instead, replace this with options that match for a specific SPI value for any IPsec protocol. An all zero SPI can now be specified to match NON-ESP NAT-T packets.

Add a funtion to the libip route class that checks if an ip address is local This is only implemented and used on windows targets.

Make sure libike and libpfk do not leak file descriptors. Rename the libpfk open and close funtions to attach and detach to make things consistent with the other IPC libraries.

Update the todo list.

git-svn-id: svn://svn.shrew.net/ike/head@199 1bb156eb-19af-db11-b919-0014c23ffdf2