Modify the iked rsa encrypt/decrypt functions to take two parameters instead of one. This is mostly just for clarity sake. Cleanup some bad code style and correct a few comments surrounding rsa authentication. Fix the xauth hybrid name strings which had the dss and rsa types reversed.
Correct a logic problem with ike payload fragmentation. If the last fragment payload size was equal to the maximum fragment size, the FRAG_FLAG_LAST bit was not being set properly. This was causing the peer to wait for additional fragments which never arrive. Cleanup some comments here as well.

git-svn-id: svn://svn.shrew.net/ike/head@334 1bb156eb-19af-db11-b919-0014c23ffdf2