Cleanup our Diffie Hellman group setup functions in iked. Don't attempt to generate a public key by hand. The OpenSSL DH_generate_key function will handle this for us. While here, improve error handling for setup failure.
Move the hash_size handler from the ike generic exchange class into the phase1 subclass. It was not being used by any other exchange subclasses.

Modify the idb member function that sets a value for a range of bytes. Specify the value parameter as an int and not a char.

git-svn-id: svn://svn.shrew.net/ike/head@521 1bb156eb-19af-db11-b919-0014c23ffdf2