Modify the idb policy class to use bitflags instead of the route and nailed boolean values. Introduce a new flag that forces a single phase2 SA to be negotiated after a policy is created. Use this option to ensure phase2 negotiation occurs immediately after a connection has been established with a Cisco gateway. These devices will disconnect a client shortly after initial contact if an IPsec SA is not established.
Modify iked to use an adaptive event timer for DHCP over IPsec. This significantly reduces the tunnel setup time when communicating with Fortinet gateways. While here, fix a bug that caused disconnects to be erroneously identified as a DHCP over IPsec configuration problem.

Modify the public version.h file to reflect a new client configuration version. Add a new definition that specifies if a build will be a stable release.

git-svn-id: svn://svn.shrew.net/ike/head@555 1bb156eb-19af-db11-b919-0014c23ffdf2