Modify iked to allow client to suspend and resume tunnel control. This is necessary for Secure Domain Login support. The virtual adapter and DNS proxy client interface instance pointers are now stored in the tunnel class instance. This allows the client control interface to be fully suspended and resumed without losing the ability to manage these subordinate interfaces. Other code paths were modified to prevent access to the client control interface pointer after a tunnel has been established. The interface pointer will be set to NULL when a client suspends tunnel control. The code that manages adapter and DNS proxy settings have been pushed into the platform specific client setup and cleanup functions. The Linux/BSD/OSX platform code will need to be updated to reflect these changes.
Modify the generic client interface to support tunnel control suspend and resume operations. These are Windows specific. The command line options parser code path was modified to not display the usage help. This was moved into its own function and should be called by clients when the option parser returns a syntax error value. The site name read by the parser is now stored internally and the load and save functions now use this instead of accepting a name value as a parameter. The Linux/BSD/OSX platform code will need to be update to reflect these changes.

git-svn-id: svn://svn.shrew.net/ike/head@796 1bb156eb-19af-db11-b919-0014c23ffdf2